I have drawable points now, but I'm running into a small issue.  It seems that the grid layout's geometry isn't matching the bounding box I'm setting.  I set a 350 x 350 view for the points to be printed in, but they only print within a small subrectangle of the grid layout...
Also, made the point list a singleton class.  No destructor or clear function yet.  To come later.

+++ trunk/Point.h
@@ -5,17 +5,17 @@ class Point
 {
 public:
     Point();
-    Point(int x, int y);
+    Point(double x, double y);
 
-    int getX();
-    int getY();
-    void setX(int x);
-    void setY(int y);
+    double getX();
+    double getY();
+    void setX(double x);
+    void setY(double y);
 
     bool operator==(const Point &p) const;
 private:
-    int x;
-    int y;
+    double x;
+    double y;
 };
 
 #endif // POINT_H
